requirements.txt
gunicorn==25.1.0
packaging==26.0
Flask
pandas
openpyxl
qrcode
pillow
nepali-datetime
moviepy
pillow